+++ cpp-terminal/iostream.cpp
@@ -28,31 +28,3 @@ TOstream&   clog = reinterpret_cast<Term::TOstream&>(clogBuf);
 static char cinBuf[sizeof(Term::TIstream)];
 TIstream&   cin = reinterpret_cast<Term::TIstream&>(cinBuf);
 }  // namespace Term */
-
-int Term::StreamInitializer::m_counter{0};
-
-void Term::StreamInitializer::init()
-{
-  if(m_counter++ == 0)
-  {
-    std::ios_base::Init();
-    new(&Term::cout) TOstream(Term::Buffer::Type::FullBuffered, BUFSIZ);
-    new(&Term::clog) TOstream(Term::Buffer::Type::LineBuffered, BUFSIZ);
-    new(&Term::cerr) TOstream(Term::Buffer::Type::Unbuffered, 0);
-    new(&Term::cin) TIstream(Term::Buffer::Type::FullBuffered, BUFSIZ);
-    if(is_stdin_a_tty()) std::cin.rdbuf(Term::cin.rdbuf());
-  }
-}
-
-Term::StreamInitializer::StreamInitializer() { init(); }
-
-Term::StreamInitializer::~StreamInitializer()
-{
-  if(--m_counter == 0)
-  {
-    (&Term::cout)->~TOstream();
-    (&Term::cerr)->~TOstream();
-    (&Term::clog)->~TOstream();
-    (&Term::cin)->~TIstream();
-  }
-}
