+++ cpp-terminal/iostream.hpp
@@ -9,6 +9,7 @@
 
 #pragma once
 
+#include "cpp-terminal/iostream_initializer.hpp"
 #include "cpp-terminal/stream.hpp"
 
 namespace Term
@@ -18,21 +19,6 @@ extern TIstream& cin;
 extern TOstream& cout;
 extern TOstream& cerr;
 extern TOstream& clog;
-
-class StreamInitializer
-{
-public:
-  StreamInitializer();
-  static void init();
-  ~StreamInitializer();
-  StreamInitializer(const StreamInitializer&)            = delete;
-  StreamInitializer& operator=(const StreamInitializer&) = delete;
-  StreamInitializer(StreamInitializer&&)                 = delete;
-  StreamInitializer& operator=(StreamInitializer&&)      = delete;
-
-private:
-  static int m_counter;
-};
 
 static StreamInitializer m_streams;
