+++ cpp-terminal/iostream_initializer.cpp
@@ -1,0 +1,44 @@
+/*
+* cpp-terminal
+* C++ library for writing multi-platform terminal applications.
+*
+* SPDX-FileCopyrightText: 2019-2023 cpp-terminal
+*
+* SPDX-License-Identifier: MIT
+*/
+
+#include "cpp-terminal/iostream_initializer.hpp"
+
+#include "cpp-terminal/iostream.hpp"
+#include "cpp-terminal/tty.hpp"
+
+#include <cstddef>
+#include <iostream>
+
+std::size_t Term::StreamInitializer::m_counter{0};
+
+void Term::StreamInitializer::init()
+{
+  if(m_counter++ == 0)
+  {
+    static const std::ios_base::Init iostreams_init;
+    new(&Term::cout) TOstream(Term::Buffer::Type::FullBuffered, BUFSIZ);
+    new(&Term::clog) TOstream(Term::Buffer::Type::LineBuffered, BUFSIZ);
+    new(&Term::cerr) TOstream(Term::Buffer::Type::Unbuffered, 0);
+    new(&Term::cin) TIstream(Term::Buffer::Type::FullBuffered, BUFSIZ);
+    if(is_stdin_a_tty()) { std::cin.rdbuf(Term::cin.rdbuf()); }
+  }
+}
+
+Term::StreamInitializer::StreamInitializer() { init(); }
+
+Term::StreamInitializer::~StreamInitializer()
+{
+  if(--m_counter == 0)
+  {
+    (&Term::cout)->~TOstream();
+    (&Term::cerr)->~TOstream();
+    (&Term::clog)->~TOstream();
+    (&Term::cin)->~TIstream();
+  }
+}
